app/services/image_service.py
# Load necessary libraries
import numpy as np
from PIL import Image

# Generate sound from image input
def generate_sound_from_image(image_url: str):
    # Load image
    image = Image.open(image_url)

    # Perform image-to-sound conversion (e.g., using models like VQ-VAE-2 or a custom approach)
    sound_data = np.random.rand(44100)  # Placeholder for actual sound data generation

    return "Generated sound from image"
