requirements.txt
fastapi
uvicorn
transformers
torch
numpy
Pillow
